fix : provider type 분리
